FIX Translate the same sentences for both models
Albeit using the same seed we cannot guarantee that exactly the same
sentences are translated for the two types of models (`m2m-100-md` and
`m2m-100-lg`), because the subset of keys is selected based on the
quantity of characters in the _target_ language (and the two models
being different may output a different number of characters for the same
input sentence). It is important to use exactly the same keys
especially in the adjusted language sampling case, because we load the
probabilities and if there is P(s|l) > 0, but there is no translation
for that pair then the code will fail, as observed by @chenQ1114!